+++ TorchSUL/Consts/Activate.py
@@ -1,0 +1,12 @@
+# activation const
+# some values are no longer supported 
+
+PARAM_RELU = 0
+PARAM_LRELU = 1
+PARAM_ELU = 2
+PARAM_TANH = 3
+PARAM_SIGMOID = 6
+PARAM_PRELU = 8
+PARAM_PRELU1 = 9
+PARAM_GELU = 10
+PARAM_SILU = 11
