+++ TorchSUL/README.md
@@ -1,0 +1,5 @@
+Prepare for version 0.3 in 2024
+
+TODO: 
+1. Support customize qbit type, observer and quantizer 
+2. Add back plotter, with better use experience
