+++ TorchSUL/Tools/Progress.py
@@ -1,0 +1,17 @@
+from rich.progress import (BarColumn, MofNCompleteColumn, Progress,
+                           ProgressColumn, TextColumn, TimeRemainingColumn)
+from rich.text import Text
+
+
+# progressbar utils 
+class SpeedColumn(ProgressColumn):
+    def render(self, task):
+        speed = task.finished_speed or task.speed
+        if speed is None:
+            return Text("?", style="progress.data.speed")
+        else:
+            return Text(f"{speed:.2f} it/s", style="progress.data.speed")
+
+def progress_bar(width: int=40) -> Progress:
+    prog = Progress(TextColumn('[progress.description]{task.description}'), BarColumn(finished_style='green', bar_width=width), MofNCompleteColumn(), TimeRemainingColumn(elapsed_when_finished=True), SpeedColumn())
+    return prog
