+++ TorchSUL/__init__.py
@@ -1,17 +1,5 @@
-from . import Layers 
-from . import Model
-from . import sul_tool
-from . import sulplotter
-from . import Base 
-from . import Quant
-from . import Config
+from .Utils import Config
+from . import Base, Model, Quant, Tools
 
-__all__ = [
-	'Layers',
-	'Model',
-	'sul_tool',
-	'sulplotter',
-	'Base',
-	'Quant',
-	'Config',
-]
+sul_tool = Tools
+
