--- TorchSUL_legacy/__init__.py
@@ -1,19 +1,0 @@
-from . import DataReader
-from . import Layers 
-from . import Model
-from . import sul_tool
-from . import sulio 
-from . import sulplotter
-from . import Base 
-from . import Quant
-
-__all__ = [
-	'DataReader',
-	'Layers',
-	'Model',
-	'sul_tool',
-	'sulio',
-	'sulplotter',
-	'Base',
-	'Quant'
-]
