--- V03/README.md
@@ -1,1 +1,0 @@
-Prepare for version 0.3 in 2024
