+++ setup.py
@@ -5,7 +5,7 @@ with open("README.md", "r") as fh:
 
 setup_args = dict(
     name='TorchSUL',
-    version='0.2.10',
+    version='0.3.0',
     description='Simple but useful layers for Pytorch',
     packages=find_packages(),
     long_description=long_description,
